Tasca 2 - Commit 5 - Test 5 passat - Cambiat el tipus de variable (int to long)

+++ Tasca2/src/Divisible3.java
@@ -6,14 +6,14 @@ public class Divisible3 {
 
     }
 
-    public String isDivisibleEntre3(int num){
+    public String isDivisibleEntre3(long num){
         String numConcatStr = "";
 
         for (int i = 0; i < num; i++) {
             numConcatStr += Integer.toString(i + 1);
         }
 
-        int numConcatenado = Integer.valueOf(numConcatStr);
+        long numConcatenado = Long.valueOf(numConcatStr);
 
         if (numConcatenado % 3 == 0){
             return "SI";
